Move fiber / plugin / service stuff around
